Revert "[WIP] #1186: Improve tests for circuit drawers (#1617)"
This reverts commit 8d3dabd2ac1388bb5cf14f8cf95ebeefc5686ac1.
