--- test/python/tools/visualization/references/large/latex_source
@@ -1,39 +1,0 @@
-% \documentclass[preview]{standalone}
-% If the image is too large to fit on this documentclass use
-\documentclass[draft]{beamer}
-% img_width = 18, img_depth = 49
-\usepackage[size=custom,height=27,width=109,scale=0.7]{beamerposter}
-% instead and customize the height and width (in cm) to fit.
-% Large images may run out of memory quickly.
-% To fix this use the LuaLaTeX compiler, which dynamically
-% allocates memory.
-\usepackage[braket, qm]{qcircuit}
-\usepackage{amsmath}
-\pdfmapfile{+sansmathaccent.map}
-% \usepackage[landscape]{geometry}
-% Comment out the above line if using the beamer documentclass.
-\begin{document}
-\begin{equation*}
-    \Qcircuit @C=0.5em @R=0.0em @!R {
-	 	\lstick{qr_{0}: \ket{0}} & \gate{X} & \gate{H} & \gate{T} & \push{\rule{.6em}{0em}\ket{0}\rule{.2em}{0em}} \qw & \gate{R_x(0.39)} & \gate{U_1(0.39)} & \qswap & \ctrl{1} & \qw & \control\qw & \ctrl{1} & \ctrl{1} & \dstick{0.39}\qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw \barrier[-1.15em]{8} & \meter & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw\\
-	 	\lstick{qr_{1}: \ket{0}} & \gate{Y} & \gate{S} & \gate{T^\dag} & \push{\rule{.6em}{0em}\ket{0}\rule{.2em}{0em}} \qw & \gate{R_y(0.39)} & \gate{U_2\left(0.39,0.39\right)} & \qswap \qwx[-1] & \targ & \ctrl{1} & \qw & \gate{H} & \control\qw & \ctrl{1} & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \meter & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw\\
-	 	\lstick{qr_{2}: \ket{0}} & \gate{Z} & \gate{S^\dag} & \qw & \push{\rule{.6em}{0em}\ket{0}\rule{.2em}{0em}} \qw & \gate{R_z(0.39)} & \gate{U_3(0.39,0.39,0.39)} & \qw & \qw & \gate{Y} & \ctrl{-2} & \qw & \qw & \gate{U_3(0.39,0.39,0.39)} & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \meter & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw\\
-	 	\lstick{qr_{3}: \ket{0}} & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \gate{X} & \gate{H} & \gate{T} & \push{\rule{.6em}{0em}\ket{0}\rule{.2em}{0em}} \qw & \gate{R_x(0.39)} & \gate{U_1(0.39)} & \qswap & \ctrl{1} & \qw & \control\qw & \ctrl{1} & \ctrl{1} & \dstick{0.39}\qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \meter & \qw & \qw & \qw & \qw & \qw & \qw & \qw\\
-	 	\lstick{qr_{4}: \ket{0}} & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \gate{Y} & \gate{S} & \gate{T^\dag} & \push{\rule{.6em}{0em}\ket{0}\rule{.2em}{0em}} \qw & \gate{R_y(0.39)} & \gate{U_2\left(0.39,0.39\right)} & \qswap \qwx[-1] & \targ & \ctrl{1} & \qw & \gate{H} & \control\qw & \ctrl{1} & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \meter & \qw & \qw & \qw & \qw & \qw & \qw\\
-	 	\lstick{qr_{5}: \ket{0}} & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \gate{Z} & \gate{S^\dag} & \qw & \push{\rule{.6em}{0em}\ket{0}\rule{.2em}{0em}} \qw & \gate{R_z(0.39)} & \gate{U_3(0.39,0.39,0.39)} & \qw & \qw & \gate{Y} & \ctrl{-2} & \qw & \qw & \gate{U_3(0.39,0.39,0.39)} & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \meter & \qw & \qw & \qw & \qw & \qw\\
-	 	\lstick{qr_{6}: \ket{0}} & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \gate{X} & \gate{H} & \gate{T} & \push{\rule{.6em}{0em}\ket{0}\rule{.2em}{0em}} \qw & \gate{R_x(0.39)} & \gate{U_1(0.39)} & \qswap & \ctrl{1} & \qw & \control\qw & \ctrl{1} & \ctrl{1} & \dstick{0.39}\qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \meter & \qw & \qw & \qw & \qw\\
-	 	\lstick{qr_{7}: \ket{0}} & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \gate{Y} & \gate{S} & \gate{T^\dag} & \push{\rule{.6em}{0em}\ket{0}\rule{.2em}{0em}} \qw & \gate{R_y(0.39)} & \gate{U_2\left(0.39,0.39\right)} & \qswap \qwx[-1] & \targ & \ctrl{1} & \qw & \gate{H} & \control\qw & \ctrl{1} & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \meter & \qw & \qw & \qw\\
-	 	\lstick{qr_{8}: \ket{0}} & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \gate{Z} & \gate{S^\dag} & \qw & \push{\rule{.6em}{0em}\ket{0}\rule{.2em}{0em}} \qw & \gate{R_z(0.39)} & \gate{U_3(0.39,0.39,0.39)} & \qw & \qw & \gate{Y} & \ctrl{-2} & \qw & \qw & \gate{U_3(0.39,0.39,0.39)} & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \qw & \meter & \qw & \qw\\
-	 	\lstick{cr_{0}: 0} & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw \cwx[-9] & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw\\
-	 	\lstick{cr_{1}: 0} & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw \cwx[-9] & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw\\
-	 	\lstick{cr_{2}: 0} & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw \cwx[-9] & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw\\
-	 	\lstick{cr_{3}: 0} & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw \cwx[-9] & \cw & \cw & \cw & \cw & \cw & \cw & \cw\\
-	 	\lstick{cr_{4}: 0} & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw \cwx[-9] & \cw & \cw & \cw & \cw & \cw & \cw\\
-	 	\lstick{cr_{5}: 0} & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw \cwx[-9] & \cw & \cw & \cw & \cw & \cw\\
-	 	\lstick{cr_{6}: 0} & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw \cwx[-9] & \cw & \cw & \cw & \cw\\
-	 	\lstick{cr_{7}: 0} & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw \cwx[-9] & \cw & \cw & \cw\\
-	 	\lstick{cr_{8}: 0} & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw & \cw \cwx[-9] & \cw & \cw\\
-	 }
-\end{equation*}
-
-\end{document}
