--- test/python/tools/visualization/references/medium/text
@@ -1,52 +1,0 @@
-         ┌───┐          ┌───┐            ┌───┐                            »
-qr_0: |0>┤ X ├──────────┤ H ├────────────┤ T ├──────────────|0>───────────»
-         └───┘┌───┐     └───┘┌───┐       └───┘┌─────┐                     »
-qr_1: |0>─────┤ Y ├──────────┤ S ├────────────┤ Tdg ├────────────|0>──────»
-              └───┘┌───┐     └───┘┌─────┐     └─────┘┌────┐               »
-qr_2: |0>──────────┤ Z ├──────────┤ Sdg ├────────────┤ Id ├───────────|0>─»
-                   └───┘          └─────┘            └────┘               »
- cr_0: 0 ═════════════════════════════════════════════════════════════════»
-                                                                          »
- cr_1: 0 ═════════════════════════════════════════════════════════════════»
-                                                                          »
- cr_2: 0 ═════════════════════════════════════════════════════════════════»
-                                                                          »
-«      ┌────────────┐                            ┌────────────┐»
-«qr_0: ┤ Rx(0.3927) ├────────────────────────────┤ U1(0.3927) ├»
-«      └────────────┘┌────────────┐              └────────────┘»
-«qr_1: ──────────────┤ Ry(0.3927) ├────────────────────────────»
-«                    └────────────┘┌────────────┐              »
-«qr_2: ────────────────────────────┤ Rz(0.3927) ├──────────────»
-«                                  └────────────┘              »
-«cr_0: ════════════════════════════════════════════════════════»
-«                                                              »
-«cr_1: ════════════════════════════════════════════════════════»
-«                                                              »
-«cr_2: ════════════════════════════════════════════════════════»
-«                                                              »
-«                                                                            »
-«qr_0: ──────────────────────────────────────────────────X───■────────■───■──»
-«      ┌───────────────────┐                             │ ┌─┴─┐      │ ┌─┴─┐»
-«qr_1: ┤ U2(0.3927,0.3927) ├─────────────────────────────X─┤ X ├──■───┼─┤ H ├»
-«      └───────────────────┘┌──────────────────────────┐   └───┘┌─┴─┐ │ └───┘»
-«qr_2: ─────────────────────┤ U3(0.3927,0.3927,0.3927) ├────────┤ Y ├─■──────»
-«                           └──────────────────────────┘        └───┘        »
-«cr_0: ══════════════════════════════════════════════════════════════════════»
-«                                                                            »
-«cr_1: ══════════════════════════════════════════════════════════════════════»
-«                                                                            »
-«cr_2: ══════════════════════════════════════════════════════════════════════»
-«                                                                            »
-«                                            ░ ┌─┐      
-«qr_0: ─■────────────────────────────────────░─┤M├──────
-«       │0.3927                              ░ └╥┘┌─┐   
-«qr_1: ─■────────────────────■───────────────░──╫─┤M├───
-«               ┌────────────┴─────────────┐ ░  ║ └╥┘┌─┐
-«qr_2: ─────────┤ U3(0.3927,0.3927,0.3927) ├─░──╫──╫─┤M├
-«               └──────────────────────────┘ ░  ║  ║ └╥┘
-«cr_0: ═════════════════════════════════════════╩══╬══╬═
-«                                                  ║  ║ 
-«cr_1: ════════════════════════════════════════════╩══╬═
-«                                                     ║ 
-«cr_2: ═══════════════════════════════════════════════╩═
-«
