--- test/python/tools/visualization/references/small/latex_source
@@ -1,23 +1,0 @@
-% \documentclass[preview]{standalone}
-% If the image is too large to fit on this documentclass use
-\documentclass[draft]{beamer}
-% img_width = 2, img_depth = 5
-\usepackage[size=custom,height=10,width=10,scale=0.7]{beamerposter}
-% instead and customize the height and width (in cm) to fit.
-% Large images may run out of memory quickly.
-% To fix this use the LuaLaTeX compiler, which dynamically
-% allocates memory.
-\usepackage[braket, qm]{qcircuit}
-\usepackage{amsmath}
-\pdfmapfile{+sansmathaccent.map}
-% \usepackage[landscape]{geometry}
-% Comment out the above line if using the beamer documentclass.
-\begin{document}
-\begin{equation*}
-    \Qcircuit @C=0.5em @R=0.0em @!R {
-	 	\lstick{qr_{0}: \ket{0}} & \gate{X} & \qw \barrier[-1.15em]{0} & \meter & \qw & \qw\\
-	 	\lstick{cr_{0}: 0} & \cw & \cw & \cw \cwx[-1] & \cw & \cw\\
-	 }
-\end{equation*}
-
-\end{document}
