--- test/python/tools/visualization/references/small/text
@@ -1,5 +1,0 @@
-         ┌───┐ ░ ┌─┐
-qr_0: |0>┤ X ├─░─┤M├
-         └───┘ ░ └╥┘
- cr_0: 0 ═════════╩═
-
